[Chore] GPTChatRoomViewController+Extension 추가
- 기존에 뷰컨트롤러에 같이 있던 익스텐션을 파일로 분리
- 뷰컨트롤러의 익스텐션에는 프로토콜 채택 내용이 닮김.
